♻️ Move collectTraitsFrom into JsonTextReplacement
Fits better with the other methods defined in that class rather than the
recursive parse methods in JsonSource
